phantomjs: use #phantomjs to get rid of #signal handling issue
When running #phantomjs as an init process, #phantomjs is not able to
handle incoming signals. For example, within
  docker run -ti --rm --name phantomjs icy/phantomjs
(to start internal command: /usr/bin/phantomjs --webdriver=8190...)
you will not be able to press ^C to stop the current #tty/phantomjs.
The only way to stop the container is to use #docker kill or #kill -9.
I do not think #docker is working right here. It is about get thing
*done* without taking care of side effects.

Here we go with #phantomjs + #supervisor. It is perfect now.